excel_import/requirement.txt
Django==2.2.6
django-excel==0.0.10
django-import-export==1.2.0
django-tablib==3.2
djongo==1.2.36
et-xmlfile==1.0.1
importlib-metadata==0.19
Jinja2==2.10.1
lxml==4.4.1
mysqlclient==1.4.4
openpyxl==2.5.14
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyexcel-xlsx==0.5.7
PyMySQL==0.9.3
xlrd==1.2.0
XlsxWriter==1.2.2
xlwt==1.3.0
